feat(power/mem): Reintroduce value-dependent icon color
Small changes that allow you to set value-dependent colors for the icons of the power and memory widget. Most of the code to make this happen was still left (as well as all parameters in the schema), but the CSS properties were missing. Added the property setting in code and added disabled template CSS code for the colors. The template colors should be consistent with the rest of the bar (Catppuccin Mocha)