chore(workflows): update Changelog action and release creation
The commit updates the Changelog action in the windows.yaml workflow to use the TriPSs/conventional-changelog-action@v5 version. It also modifies the release creation step to only execute if the changelog step was not skipped. This ensures that the release is created only when there are actual changes to include in the changelog.