feat(widget): Libre Hardware Monitor Widget
This commit introduces the Libre Hardware Monitor widget that connects to the local LHM web server and fetches the selected sensor data using SensorId.
The post request is done asynchronously using QNetworkAccessManager to avoid locking the UI.
Possible errors are handled and will be displayed in the widget instead of the value.
Auth is supported.
Histogram view is also available.